pyproject.toml
[project]
name = "medical-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
